+++ src/main/java/com/example/thisaraprinters/repository/CustomerRepo.java
@@ -3,13 +3,13 @@ package com.example.thisaraprinters.repository;
 import com.example.thisaraprinters.model.CustomerModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepo extends JpaRepository<CustomerModel, Long> {
+public interface CustomerRepo extends JpaRepository<CustomerModel, Integer> {
     boolean existsByName(String name);
     boolean existsByEmail(String email);
     boolean existsByPhone(String phone);
 
-    boolean existsByNameAndIdNot(String name, Long id);
-    boolean existsByEmailAndIdNot(String email, Long id);
-    boolean existsByPhoneAndIdNot(String phone, Long id);
+    boolean existsByNameAndIdNot(String name, Integer id);
+    boolean existsByEmailAndIdNot(String email, Integer id);
+    boolean existsByPhoneAndIdNot(String phone, Integer id);
 }
